Removed Services, and such, Got Texture rendering, no state atm

+++ L12.Patterns/Direct3D.inl
@@ -9,8 +9,8 @@ GraphicsDevice::InputLayout GraphicsDevice::CreateInputLayout(const std::array<D
 {
 	InputLayout il;
 
-	HRESULT hr = m_Device->CreateInputLayout(elements.data(), elements.size(),
-		vsByteCode.data(), vsByteCode.size(),
+	HRESULT hr = m_Device->CreateInputLayout(elements.data(), (uint32_t)elements.size(),
+		vsByteCode.data(), (uint32_t)vsByteCode.size(),
 		&il);
 	ThrowIfFailed(hr, "Failed to create input layout");
 
@@ -20,8 +20,8 @@ GraphicsDevice::InputLayout GraphicsDevice::CreateInputLayout(const std::array<D
 template<uint16_t SIZE>
 bool GraphicsDevice::CheckInputLayout(const std::array<D3D11_INPUT_ELEMENT_DESC, SIZE>& elements, const std::vector<byte>& vsByteCode)
 {
-	HRESULT hr = m_Device->CreateInputLayout(elements.data(), elements.size(),
-		vsByteCode.data(), vsByteCode.size(),
+	HRESULT hr = m_Device->CreateInputLayout(elements.data(), (uint32_t)elements.size(),
+		vsByteCode.data(), (uint32_t)vsByteCode.size(),
 		NULL);
 	// CreateInputLayout returns S_FALSE, 
 	// if the signature of the InputLayout Matches that of the Shader's Inputs
